+++ scripts/data/packs.js
@@ -14,6 +14,8 @@ import { SURROUNDED_REFLEX_CARDS } from "./cards/surrounded-reflex.js";
 import { SURROUNDED_WILL_CARDS } from "./cards/surrounded-will.js";
 import { GIANT_SLAYER_ATTACK_CARDS } from "./cards/giant-slayer-attack.js";
 import { GIANT_SLAYER_FORTITUDE_CARDS } from "./cards/giant-slayer-fortitude.js";
+import { GIANT_SLAYER_REFLEX_CARDS } from "./cards/giant-slayer-reflex.js";
+import { GIANT_SLAYER_WILL_CARDS } from "./cards/giant-slayer-will.js";
 
 const THEMES = Object.freeze([
   Object.freeze({
@@ -89,7 +91,11 @@ export function buildAgainstAllOddsPacks(settings = {}) {
                         ? GIANT_SLAYER_ATTACK_CARDS
                         : theme.id === THEME_IDS.GIANT_SLAYER && deckType === "fortitude"
                           ? GIANT_SLAYER_FORTITUDE_CARDS
-                          : []
+                          : theme.id === THEME_IDS.GIANT_SLAYER && deckType === "reflex"
+                            ? GIANT_SLAYER_REFLEX_CARDS
+                            : theme.id === THEME_IDS.GIANT_SLAYER && deckType === "will"
+                              ? GIANT_SLAYER_WILL_CARDS
+                              : []
     }]))
   }));
 }